requirements.txt
bcrypt==3.2.0
blinker==1.4
boto==2.49.0
boto3==1.17.50
botocore==1.20.50
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
dnspython==2.1.0
email-validator==1.1.2
firebase==3.0.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gcloud==0.17.0
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
packaging==20.9
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
SQLAlchemy==1.4.8
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
